Update size_t to int

+++ 7. C# DataStructures/1.DataStructuresFundamentals/04. Linear Data Structures/C++/LinearDataStructures/Problem01.List/source/list/IAbstractList.h
@@ -2,7 +2,7 @@
 #ifndef IABSTRACTLIST_H
 #define IABSTRACTLIST_H
 
-#include <cstddef>
+//#include <cstddef>
 
 template<typename T>
 class IAbstractList
@@ -10,16 +10,16 @@ class IAbstractList
 public:
 	virtual ~IAbstractList() = default;
 
-	virtual size_t Count() const = 0;
-	virtual T& operator[](size_t index) = 0;
-	virtual const T& operator(size_t index) const = 0;
+	virtual int Count() const = 0;
+	virtual T& operator[](int index) = 0;
+	virtual const T& operator(int index) const = 0;
 
 	virtual void Add(const T& item) = 0;
-	virtual void Insert(size_t index, const T& item) = 0;
+	virtual void Insert(int index, const T& item) = 0;
 	virtual bool Contains(const T& item) = 0;
 	virtual int IndexOf(const T& item) = 0;
 	virtual bool Remove(const T& item) = 0;
-	virtual void RemoveAt(size_t index) = 0;
+	virtual void RemoveAt(int index) = 0;
 };
 
 #endif
